RDM-2814 - review comments - used alternative rest api to delete index alias
