+++ aat/src/aat/java/uk/gov/hmcts/ccd/definitionstore/tests/functional/elasticsearch/ElasticsearchImportDefinitionTest.java
@@ -4,25 +4,19 @@ import java.io.File;
 import java.util.stream.Stream;
 
 import static java.util.Optional.ofNullable;
-import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.not;
 import static org.hamcrest.Matchers.hasKey;
 import static org.junit.jupiter.api.Assumptions.assumeTrue;
 
-import io.restassured.RestAssured;
-import io.restassured.builder.RequestSpecBuilder;
-import io.restassured.http.ContentType;
 import io.restassured.response.ValidatableResponse;
-import io.restassured.specification.RequestSpecification;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import uk.gov.hmcts.ccd.definitionstore.tests.AATHelper;
-import uk.gov.hmcts.ccd.definitionstore.tests.BaseTest;
 
-class ElasticsearchImportDefinitionTest extends BaseTest {
+class ElasticsearchImportDefinitionTest extends ElasticsearchBaseTest {
 
     private static final String DEFINITION_FILE = "src/resource/CCD_CNP_27.xlsx";
     private static final String DEFINITION_FILE_WITH_NEW_FIELD = "src/resource/CCD_CNP_27_WithNewField.xlsx";
@@ -37,7 +31,7 @@ class ElasticsearchImportDefinitionTest extends BaseTest {
 
     @BeforeAll
     static void setUp() {
-        // stop execution of these tests if the env variable is not set
+        // stop execution of these tests if elasticsearch is not enabled
         boolean elasticsearchEnabled = ofNullable(System.getenv("ELASTIC_SEARCH_ENABLED")).map(Boolean::valueOf).orElse(false);
         assumeTrue(elasticsearchEnabled, () -> "Ignoring Elasticsearch tests, variable ELASTIC_SEARCH_ENABLED not set");
     }
@@ -70,46 +64,13 @@ class ElasticsearchImportDefinitionTest extends BaseTest {
 
     @AfterEach
     void tearDown() {
-        removeAlias();
-        deleteIndex();
-    }
-
-    private void removeAlias() {
-        asElasticsearchApiUser()
-            .given()
-            .contentType(ContentType.JSON)
-            .body("{\"actions\":[{\"remove\":{\"index\":\"" + CASE_INDEX_NAME + "\",\"alias\":\"" + CASE_INDEX_ALIAS + "\"}}]}")
-            .when()
-            .post("_aliases")
-            .then()
-            .statusCode(200);
-    }
-
-    private void deleteIndex() {
-        // deletes the index
-        asElasticsearchApiUser()
-            .when()
-            .delete(CASE_INDEX_NAME)
-            .then()
-            .statusCode(200)
-            .body("acknowledged", equalTo(true));
+        deleteIndex(CASE_INDEX_NAME, CASE_INDEX_ALIAS);
     }
 
     private void verifyIndexAndFieldMappings(boolean verifyNewFields) {
-        ValidatableResponse response = asElasticsearchApiUser()
-            .when()
-            .get(CASE_INDEX_ALIAS)
-            .then()
-            .statusCode(200);
-
+        ValidatableResponse response = getIndexInformation(CASE_INDEX_ALIAS);
         verifyIndexAndAlias(response);
         verifyCaseDataFields(response, verifyNewFields);
-    }
-
-    private RequestSpecification asElasticsearchApiUser() {
-        return RestAssured.given(new RequestSpecBuilder()
-                                     .setBaseUri(aat.getElasticsearchBaseUri())
-                                     .build());
     }
 
     private void verifyIndexAndAlias(ValidatableResponse response) {
